made hashutils
